refactor: extract kiro_streaming.go and kiro_transform.go from kiro_executor.go
Split the 4691-line kiro_executor.go into three focused files:

- kiro_transform.go (~470 LOC): endpoint config types, region resolution,
  payload builders (buildKiroPayloadForFormat, sanitizeKiroPayload),
  model mapping (mapModelToKiro), credential extraction (kiroCredentials),
  and auth-method helpers (getEffectiveProfileArnWithWarning, isIDCAuth).

- kiro_streaming.go (~2990 LOC): streaming execution (ExecuteStream,
  executeStreamWithRetry), AWS Event Stream parsing (parseEventStream,
  readEventStreamMessage, extractEventTypeFromBytes), channel-based
  streaming (streamToChannel), and the full web search MCP handler
  (handleWebSearchStream, handleWebSearch, callMcpAPI, etc.).

- kiro_executor.go (~1270 LOC): core executor struct (KiroExecutor),
  HTTP client pool, retry logic, Execute/executeWithRetry,
  CountTokens, Refresh, and token persistence helpers.

All functions remain in the same package; no public API changes.
